Move handler code into seprate module in hyp3_process
