+++ modules/hyp3_process/src/hyp3_process/handler/__init__.py
@@ -1,0 +1,5 @@
+from .handler import (
+    make_hyp3_processing_function_from,
+    HandlerFunction,
+    ProcessingFunction
+)
