+++ modules/hyp3_process/src/hyp3_process/handler/handler.py
@@ -8,7 +8,7 @@ from . import working_directory
 from .outputs import OutputPatterns
 from . import package
 from . import products
-from .hyp3_daemon import log
+from ..hyp3_daemon import log
 
 
 HandlerFunction = Callable[[
